필요 없는 Interface 및 클래스 삭제
Ladder 인터페이스나 LadderDrawerByPeopleCount클래스가 필요없다고 판단하여 삭제
네이밍
LadderDrawerByNames -> LadderDrawer
